Extract pipe
