+++ web-app/src/games/api.admin.games.controller.ts
@@ -10,7 +10,7 @@ import {
 import { AuthGuard } from '@nestjs/passport';
 import { Roles } from 'src/auth/roles.decorator';
 import { RolesGuard } from 'src/auth/roles.guard';
-import { SimpleJsonParsePipe } from './api.admin.games.pipes';
+import { SimpleJsonParsePipe } from '../nest-common/simpleJsonParse.pipes';
 import {
   ApiAdminGamesService,
   GamesSearchParams,
