+++ web-app/src/nest-common/simpleJsonParse.pipes.ts
